Backend/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
pillow
